requirements.txt
langchain
openai
faiss-cpu
pypdf
tiktoken
python-dotenv
